refactor(pending): rename pending context module and update imports
- Renamed `src/context/pending.tsx` to `src/context/pendingMessage.tsx`
